pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "python-eyelinkparser"
description = "Parsing EyeLink data in Python"
version = "0.18.0"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["eye-tracking", "EyeLink"]
license = {text = "GNU GENERAL PUBLIC LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy>=1.18",
    "scipy>=1.7",
    "python-datamatrix>=0.13",
    "fastnumbers>=3.0",
    "prettytable>=3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["python_eyelinkparser"]